dat/opthelp formatting (mostly)
Make the formatting of 'opthelp' be more consistent.  When entries
span multiple lines, sometimes the default value was shown on the
first line, sometimes on the last.  This puts it on the first line
which is more useful for quick looks with 'grep'.

This also adds an explanation for the value of 'disclose', adds
missing choice "full compass" to 'whatis_coord', and expands upon
the explanation for several other compound options.

As mentioned at least once before, menu_* is missing win32 in the
list of interfaces that support the various menu accelerators.