doc/tmac.n: UNIX trademark
The 'roff Guidebook's first reference to UNIX includes a footnote with
the trademark notice.  Several subsequent ones do not.  I think that's
ok; I think the '.ux' macro even keeps track and only issues it once.
But the extra ones aren't currently using '.ux' and probably should in
order to get the '(tm)' character.  I'll leave that for someone else....

The history section mentions quite a few trademarked names.  I hope
that the blanket "brand and product names are trademarks or registered
trademarks of their respective holders" is sufficient for them.